Revert "[feature] Ascend NPU graph support (#8027)" (#9348)
